chore(quality): make qypr strict-tidy clean
Strict clang-tidy (--warnings-as-errors='*') was red on ~2100 findings
tree-wide, so the tidy gate failed on files no series touched. Per rule 3,
pedantic checks firing on idiomatic code are disabled in .clang-tidy with
reasons (mirroring waylaunch precedent); legitimate findings (C-array,
malloc, empty-catch, uninitialized, missing-default, naming) are fixed in
code with behavior-preserving edits.

+++ qypr/src/core/App.cpp
@@ -193,7 +193,7 @@ int App::videoTest(int seconds) {
     loop_.addTimer(1000, true, [this] {
         std::fprintf(stderr, "video-test: hasFrame=%d\n", video_.hasFrame() ? 1 : 0);
     });
-    loop_.addTimer(seconds * 1000, false, [this] { loop_.quit(); });
+    loop_.addTimer(static_cast<int64_t>(seconds) * 1000, false, [this] { loop_.quit(); });
 
     loop_.run();
     std::fprintf(stderr, "video-test: DONE hasFrame=%d\n", video_.hasFrame() ? 1 : 0);
